feat: web 앱 Next.js, React 버전 다운그레이드
