+++ apps/web/.gitignore
@@ -25,8 +25,8 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-# env files (can opt-in for commiting if needed)
-.env*
+# local env files
+.env*.local
 
 # vercel
 .vercel
